refactor: 수익 계산 로직
